+++ Assets/WorldGeneration/_Scripts/Spawning/AssetManager.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using UnityEngine;
 using WorldGeneration._Scripts.Helper;
 using WorldGeneration._Scripts.ScriptableObjects;
@@ -9,19 +10,20 @@ namespace WorldGeneration._Scripts.Spawning
     [Serializable]
     public class AssetManager
     {
+        public static AssetManager Instance;
+        
         // private
-        private static AssetManager _instance;
-        private AssetSpawner _assetSpawner;
+        private List<AssetSettings> assets;
 
         public static AssetManager GetInstance()
         {
-            if (_instance == null)
+            if (Instance == null)
             {
-                _instance = new AssetManager();
-                return _instance;
+                Instance = new AssetManager();
+                return Instance;
             }
-
-            return _instance;
+        
+            return Instance;
         }
 
         /// <summary>
@@ -30,28 +32,18 @@ namespace WorldGeneration._Scripts.Spawning
         /// <param name="maxTerrainHeight"></param>
         /// <param name="generalSettings"></param>
         /// <param name="map"></param>
-        /// <param name="carrotSettings"></param>
-        /// <param name="carrots"></param>
-        /// <param name="grassSettings"></param>
-        /// <param name="grass"></param>
-        public void SpawnAssets(Vector2Int resolution, float maxTerrainHeight, GeneralSettings generalSettings,
-            float[,] map, PlantSettings carrotSettings, Transform carrots, PlantSettings grassSettings, Transform grass)
+        /// <param name="settings"></param>
+        /// <param name="parent"></param>
+        public void InitialSpawnAssets(Vector2Int resolution, float maxTerrainHeight, GeneralSettings generalSettings,
+            float[,] map, AssetSettings settings, Transform parent)
         {
-            _assetSpawner = new AssetSpawner();
+            assets = new List<AssetSettings>();
 
-            for (var i = 0; i < carrotSettings.maxNumber; i++)
+            for (var i = 0; i < settings.maxNumber; i++)
             {
                 // var posFound = false;
                 float xPos, zPos;
                 Vector3 pos;
-
-                // xPos = 3.346446854546165441312716954234876589f;
-                // zPos = 6.023497893456012363497379756109274570f;
-                //
-                // pos = GeneratorFunctions.GetSurfacePoint(xPos, zPos, resolution, maxTerrainHeight, generalSettings,
-                //     noiseSettings, noiseWithClamp);
-                //
-                // Debug.Log(pos.ToString());
 
                 // while (!posFound)
                 // {
@@ -59,7 +51,7 @@ namespace WorldGeneration._Scripts.Spawning
                 //     generalSettings.squareSize * (resolution.x - 1) / 2 + generalSettings.squareSize * (resolution.x - 1) / 2);
                 // zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
                 //     generalSettings.squareSize * (resolution.y - 1) / 2 + generalSettings.squareSize * (resolution.y - 1) / 2);
-                
+
                 xPos = Random.Range(-generalSettings.squareSize * (resolution.x - 1) / 2,
                     generalSettings.squareSize * (resolution.x - 1) / 2);
                 zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
@@ -70,16 +62,16 @@ namespace WorldGeneration._Scripts.Spawning
                 // Debug.Log("zPos: "+ zPos);
 
                 pos = GeneratorFunctions.GetSurfacePointFromWorldCoordinate(xPos, zPos, resolution, maxTerrainHeight,
-                    map,
-                    generalSettings);
+                    map, generalSettings);
                 // pos = GeneratorFunctions.GetSurfacePoint(xPos, zPos, resolution, maxTerrainHeight, map,
                 //     generalSettings);
                 // if (pos.y > maxTerrainHeight * waterLevel && pos.y < maxTerrainHeight * 0.8f)
                 // {
-                var carrot = _assetSpawner.Spawn(carrotSettings.plantPrefab, pos);
-                carrot.transform.SetParent(carrots);
+                var asset = UnityEngine.Object.Instantiate(settings.assetPrefab, pos, Quaternion.Euler(new Vector3(0, Random.Range(0, 360), 0)));
+                asset.transform.SetParent(parent);
 
-                carrotSettings.plants.Add(carrot);
+                settings.assets.Add(asset);
+                assets.Add(settings);
 
                 // posFound = true;
                 // }
@@ -93,32 +85,66 @@ namespace WorldGeneration._Scripts.Spawning
                 //         generalSettings, noiseSettings, noiseWithClamp));
             }
 
-            for (var i = 0; i < grassSettings.maxNumber; i++)
-            {
-                // var posFound = false;
-                float xPos, zPos;
-                Vector3 pos;
+            // for (var i = 0; i < grassSettings.maxNumber; i++)
+            // {
+            //     // var posFound = false;
+            //     float xPos, zPos;
+            //     Vector3 pos;
+            //
+            //     xPos = Random.Range(-generalSettings.squareSize * (resolution.x - 1) / 2,
+            //         generalSettings.squareSize * (resolution.x - 1) / 2);
+            //     zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
+            //         generalSettings.squareSize * (resolution.y - 1) / 2);
+            //
+            //     // xPos = Random.Range(
+            //     //     -generalSettings.squareSize * resolution.x / 2 /* + generalSettings.squareSize / 2*/,
+            //     //     -generalSettings.squareSize * resolution.x / 2 + generalSettings.squareSize * resolution.x /*+
+            //     //     generalSettings.squareSize / 2*/);
+            //     // zPos = Random.Range(
+            //     //     -generalSettings.squareSize * resolution.y / 2 /* + generalSettings.squareSize / 2*/,
+            //     //     -generalSettings.squareSize * resolution.y / 2 + generalSettings.squareSize * resolution.y /*+
+            //     //     generalSettings.squareSize / 2*/);
+            //
+            //     pos = GeneratorFunctions.GetSurfacePointFromWorldCoordinate(xPos, zPos, resolution, maxTerrainHeight,
+            //         map, generalSettings);
+            //
+            //     var plant = _assetSpawner.Spawn(grassSettings.plantPrefab, pos);
+            //     plant.transform.SetParent(grassSettings.parent);
+            // }
+        }
 
-                xPos = Random.Range(-generalSettings.squareSize * (resolution.x - 1) / 2,
-                    generalSettings.squareSize * (resolution.x - 1) / 2);
-                zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
-                    generalSettings.squareSize * (resolution.y - 1) / 2);
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="resolution"></param>
+        /// <param name="maxTerrainHeight"></param>
+        /// <param name="generalSettings"></param>
+        /// <param name="map"></param>
+        /// <param name="settings"></param>
+        /// <param name="parent"></param>
+        /// <param name="count"></param>
+        public void SpawnAssets(Vector2Int resolution, float maxTerrainHeight, GeneralSettings generalSettings,
+            float[,] map, AssetSettings settings, Transform parent, int count)
+        {
+                for (var i = 0; i < count; i++)
+                {
+                    // var posFound = false;
+                    float xPos, zPos;
+                    Vector3 pos;
+                    
+                    xPos = Random.Range(-generalSettings.squareSize * (resolution.x - 1) / 2,
+                        generalSettings.squareSize * (resolution.x - 1) / 2);
+                    zPos = Random.Range(-generalSettings.squareSize * (resolution.y - 1) / 2,
+                        generalSettings.squareSize * (resolution.y - 1) / 2);
 
-                // xPos = Random.Range(
-                //     -generalSettings.squareSize * resolution.x / 2 /* + generalSettings.squareSize / 2*/,
-                //     -generalSettings.squareSize * resolution.x / 2 + generalSettings.squareSize * resolution.x /*+
-                //     generalSettings.squareSize / 2*/);
-                // zPos = Random.Range(
-                //     -generalSettings.squareSize * resolution.y / 2 /* + generalSettings.squareSize / 2*/,
-                //     -generalSettings.squareSize * resolution.y / 2 + generalSettings.squareSize * resolution.y /*+
-                //     generalSettings.squareSize / 2*/);
+                    pos = GeneratorFunctions.GetSurfacePointFromWorldCoordinate(xPos, zPos, resolution,
+                        maxTerrainHeight, map, generalSettings);
 
-                pos = GeneratorFunctions.GetSurfacePointFromWorldCoordinate(xPos, zPos, resolution, maxTerrainHeight,
-                    map, generalSettings);
+                    var carrot = UnityEngine.Object.Instantiate(settings.assetPrefab, pos, Quaternion.Euler(new Vector3(0, Random.Range(0, 360), 0)));
+                    carrot.transform.SetParent(parent);
 
-                var plant = _assetSpawner.Spawn(grassSettings.plantPrefab, pos);
-                plant.transform.SetParent(grass);
-            }
+                    settings.assets.Add(carrot);
+                }
         }
     }
 }