+++ Assets/WorldGeneration/_Scripts/Spawning/TerrainAssets/RabbitBurrow.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using InGameTime;
 using UnityEngine;
 using WorldGeneration._Scripts.ScriptableObjects;
@@ -9,6 +10,8 @@ namespace WorldGeneration._Scripts.Spawning.TerrainAssets
         [SerializeField] private BurrowSettings settings;
         private InGameDate _birthDate;
         private TimeManager _timer;
+        private bool _canEnter;
+        private List<GameObject> _inhabitants;
 
         private void Start()
         {
@@ -21,6 +24,31 @@ namespace WorldGeneration._Scripts.Spawning.TerrainAssets
         private void Update()
         {
             if (_birthDate.AddDates(settings.lifespan).Equals(_timer.GetCurrentDate())) Destroy(gameObject);
+        }
+        
+        public bool Interact(GameObject interactingObject)
+        {
+            if (interactingObject.GetComponent<CustomAgent>().type == AgentType.FOX)
+            {
+                _canEnter = false;
+            }
+            else if (interactingObject.GetComponent<CustomAgent>().type == AgentType.RABBIT)
+            {
+                if (_inhabitants.Count < 2)
+                {
+                    interactingObject.GetComponent<CustomAgent>().canMove = false;
+                    _inhabitants.Add(interactingObject);
+
+                    _canEnter = true;
+                }
+                else
+                {
+                    _canEnter = false;
+                }
+                
+            }
+
+            return _canEnter;
         }
     }
 }